+++ packages/twenty-server/src/engine/api/graphql/graphql-query-runner/enums/database-event-action.ts
@@ -1,0 +1,6 @@
+export enum DatabaseEventAction {
+  CREATED = 'created',
+  UPDATED = 'updated',
+  DELETED = 'deleted',
+  DESTROYED = 'destroyed',
+}
