+++ packages/twenty-server/src/engine/api/graphql/graphql-query-runner/utils/check-string-is-database-event-action.ts
@@ -1,4 +1,4 @@
-import { DatabaseEventAction } from 'src/engine/api/graphql/graphql-query-runner/services/api-event-emitter.service';
+import { DatabaseEventAction } from 'src/engine/api/graphql/graphql-query-runner/enums/database-event-action';
 
 export const checkStringIsDatabaseEventAction = (value: string): boolean => {
   return Object.values(DatabaseEventAction).includes(
