+++ packages/twenty-server/src/engine/core-modules/open-api/utils/computeWebhooks.utils.ts
@@ -2,7 +2,7 @@ import { OpenAPIV3_1 } from 'openapi-types';
 
 import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
 import { capitalize } from 'src/utils/capitalize';
-import { DatabaseEventAction } from 'src/engine/api/graphql/graphql-query-runner/services/api-event-emitter.service';
+import { DatabaseEventAction } from 'src/engine/api/graphql/graphql-query-runner/enums/database-event-action';
 
 export const computeWebhooks = (
   type: DatabaseEventAction,
